requirements.txt
# Sightengine Text Moderation Service - Dependencies

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Optional: untuk type checking dan development
# mypy>=1.0.0
# black>=23.0.0
# pytest>=7.0.0